requirements.txt
torch~=1.10.1
torchvision~=0.11.2
pyyaml~=6.0
pillow~=8.4.0
numpy~=1.19.5
matplotlib~=3.3.4
scikit-learn~=0.24.2
pyqtgraph~=0.11.1
pyqt5~=5.15.6
scipy~=1.5.4
seaborn
opencv-python-headless
colour-demosaicing